Extract Tron into walletkit-chain-tron module
Move the Tron adapters, kit manager, account manager, send flow,
TRC20 services and TronChainPlugin into their own Gradle module and
move the tron-kit dependency with them. Walletkit core no longer
depends on any blockchain SDK; an app that omits the module ships no
trace of Tron.
