+++ settings.gradle.kts
@@ -26,6 +26,7 @@ include(":walletkit-chain-zcash")
 include(":walletkit-chain-solana")
 include(":walletkit-chain-stellar")
 include(":walletkit-chain-ton")
+include(":walletkit-chain-tron")
 include(":walletkit-chain-thorchain")
 include(":walletkit-chain-bitcoin")
 include(":walletkit-chain-evm")
