+++ walletkit-chain-tron/src/main/java/io/horizontalsystems/walletkit/core/adapters/TronAdapter.kt
