+++ walletkit/src/main/java/io/horizontalsystems/walletkit/core/managers/AccountCleaner.kt
@@ -2,7 +2,6 @@ package io.horizontalsystems.walletkit.core.managers
 
 import io.horizontalsystems.walletkit.core.IAccountCleaner
 import io.horizontalsystems.walletkit.core.chain.ChainRegistry
-import io.horizontalsystems.walletkit.core.adapters.TronAdapter
 
 class AccountCleaner : IAccountCleaner {
 
@@ -11,7 +10,6 @@ class AccountCleaner : IAccountCleaner {
     }
 
     private fun clearAccount(accountId: String) {
-        TronAdapter.clear(accountId)
         ChainRegistry.all.forEach { it.clearAccountData(accountId) }
     }
